ForYouTheStoreIsClosed
When a project targeted both net45 and wpa81 or win8 we did not have a subset that could satisfy this. The store websocket transport can only be used if an app/lib targets wpa81 and/or win8 since it is using WinRT APIs. If the app/lib targets any other platform in addition to wpa81/win8 the store websocket cannot be used and therefore we must not install it. Adding wpa81/win8 platforms to the folder containing just the portable version of Microsoft.AspNet.SignalR.Client.dll will do the trick - this folder will be selected if the application targets any other platform in addition to wpa81/win8

Fixes #3393